Extract artefact filter into reusable module
Move finding-specific filter logic so it can be shared by findings_cfg
and extensions_cfg
